+++ assets/src/user-panel.js
@@ -20,6 +20,8 @@ function UserPanel({ config }) {
   const [editBooking, setEditBooking] = useState(null);
   const [deleteBooking, setDeleteBooking] = useState(null);
   const [changedBookingId, setChangedBookingId] = useState(null);
+  const [sortField, setSortField] = useState("start");
+  const [sortDirection, setSortDirection] = useState("desc");
 
   const onBookingChanged = (bookingId, start, end) => {
     setEditBooking(null);
@@ -94,6 +96,59 @@ function UserPanel({ config }) {
     }
   };
 
+  const sortedBookings = (bookings, field, direction) => {
+    let result = bookings.sort((a, b) => {
+      return a[field] - b[field];
+    });
+
+    if (direction === "desc") {
+      result = result.reverse();
+    }
+
+    return result;
+  };
+
+  const setSort = (field) => {
+    if (sortField === field) {
+      setSortDirection(sortDirection === "asc" ? "desc" : "asc");
+    } else {
+      setSortDirection("desc");
+
+      setSortField(field);
+    }
+  };
+
+  const renderSortingButton = (field, title) => {
+    return (
+      <button
+        type="button"
+        onClick={() => setSort(field)}
+        className={`userbookings-sorting${sortField === field ? " active" : ""}`}
+      >
+        {sortField === field && (sortDirection === "asc" ? "↑ " : "↓ ")}
+        {title}
+      </button>
+    );
+  };
+
+  const renderBooking = (booking) => {
+    return (
+      <>
+        <div>
+          {booking.id === changedBookingId && <>Ændring gennemført.</>}
+          <span className="location">{booking.displayName}</span>
+          <span className="subject">{booking.subject}</span>
+          <span className="status">{getStatus(booking.status)}</span>
+        </div>
+        <div>
+          <span>{getFormattedDateTime(booking.start)}</span>
+          <span>→</span>
+          <span>{getFormattedDateTime(booking.end)}</span>
+        </div>
+      </>
+    );
+  };
+
   return (
     <div className="App">
       <div className="container-fluid">
@@ -112,25 +167,22 @@ function UserPanel({ config }) {
           )}
           {!editBooking && !deleteBooking && (
             <div className="userpanel row">
-              <div className="col no-gutter">
+              <div className="col no-padding">
                 {loading && <LoadingSpinner />}
+                {!loading && !editBooking && (
+                  <div className="userbookings-sorting-container">
+                    {renderSortingButton("displayName", "Lokale/Resurse")}
+                    {renderSortingButton("start", "Dato")}
+                    {renderSortingButton("subject", "Titel")}
+                  </div>
+                )}
                 {!loading && !editBooking && userBookings && (
                   <div className="userbookings-container">
                     <h3>Aktive bookinger</h3>
 
-                    {currentBookings.map((obj) => (
+                    {sortedBookings(currentBookings, sortField, sortDirection).map((obj) => (
                       <div className="user-booking" key={obj.id}>
-                        <div>
-                          {obj.id === changedBookingId && <>Ændring gennemført.</>}
-                          <span className="location">{obj.resourceDisplayName}</span>
-                          <span className="subject">{obj.subject}</span>
-                          <span className="status">{getStatus(obj.status)}</span>
-                        </div>
-                        <div>
-                          <span>{getFormattedDateTime(obj.start)}</span>
-                          <span>→</span>
-                          <span>{getFormattedDateTime(obj.end)}</span>
-                        </div>
+                        {renderBooking(obj)}
                         <div>
                           <button type="button" onClick={() => setDeleteBooking(obj)}>
                             Anmod om sletning
@@ -147,19 +199,9 @@ function UserPanel({ config }) {
                 {!loading && !editBooking && expiredBookings && (
                   <div className="userbookings-container">
                     <h3>Afsluttede bookinger</h3>
-                    {expiredBookings.map((obj) => (
+                    {sortedBookings(expiredBookings, sortField, sortDirection).map((obj) => (
                       <div className="user-booking expired" key={obj.id}>
-                        <div>
-                          {obj.id === changedBookingId && <>Ændring gennemført.</>}
-                          <span className="location">{obj.resourceDisplayName}</span>
-                          <span className="subject">{obj.subject}</span>
-                          <span className="status">{getStatus(obj.status)}</span>
-                        </div>
-                        <div>
-                          <span>{getFormattedDateTime(obj.start)}</span>
-                          <span>→</span>
-                          <span>{getFormattedDateTime(obj.end)}</span>
-                        </div>
+                        {renderBooking(obj)}
                       </div>
                     ))}
                   </div>