+++ CHANGELOG.md
@@ -11,6 +11,7 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 - Use resourceDisplayName when displaying resources to the user.
 - Remove html from resourceDescription
 - Change map to use locationDisplayName for filter label and locationName as filter value.
+- Sorting to user panel.
 
 ### Changed